+++ platform/engines/ospray/ispc/camera/PerspectiveStereoCamera.cpp
@@ -51,15 +51,16 @@ void PerspectiveStereoCamera::commit()
     apertureRadius = getParamf(CAMERA_PROPERTY_APERTURE_RADIUS.name.c_str(), DEFAULT_CAMERA_APERTURE_RADIUS);
     focalDistance = getParamf(CAMERA_PROPERTY_FOCAL_DISTANCE.name.c_str(), DEFAULT_CAMERA_FOCAL_DISTANCE);
     nearClip = getParamf(CAMERA_PROPERTY_NEAR_CLIP.name.c_str(), DEFAULT_CAMERA_NEAR_CLIP);
-    stereoMode = getParam(CAMERA_PROPERTY_STEREO.name.c_str(), DEFAULT_CAMERA_STEREO) ? CameraStereoMode::side_by_side
-                                                                                      : CameraStereoMode::mono;
+    stereoMode = getParam(CAMERA_PROPERTY_STEREO.name.c_str(), static_cast<int>(DEFAULT_CAMERA_STEREO))
+                     ? CameraStereoMode::side_by_side
+                     : CameraStereoMode::mono;
     interpupillaryDistance =
         getParamf(CAMERA_PROPERTY_INTERPUPILLARY_DISTANCE.name.c_str(), DEFAULT_CAMERA_INTERPUPILLARY_DISTANCE);
     enableClippingPlanes =
         getParam(CAMERA_PROPERTY_ENABLE_CLIPPING_PLANES.name.c_str(), DEFAULT_CAMERA_ENABLE_CLIPPING_PLANES);
     clipPlanes = enableClippingPlanes ? getParamData(CAMERA_PROPERTY_CLIPPING_PLANES, nullptr) : nullptr;
-    useHardwareRandomizer =
-        getParam(COMMON_PROPERTY_USE_HARDWARE_RANDOMIZER.name.c_str(), DEFAULT_COMMON_USE_HARDWARE_RANDOMIZER);
+    useHardwareRandomizer = getParam(COMMON_PROPERTY_USE_HARDWARE_RANDOMIZER.name.c_str(),
+                                     static_cast<int>(DEFAULT_COMMON_USE_HARDWARE_RANDOMIZER));
 
     dir = normalize(dir);
     vec3f dir_du = normalize(cross(dir, up));
